% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGEMatrix.R
\name{addTrt}
\alias{addTrt}
\title{Add treatment}
\arguments{
\item{x}{A \code{character}. The name of an expression that has been
downloaded from the connection.}
}
\value{
An \code{ExpressionSet} with new phenoData.
}
\description{
Add treatment information to the phenoData of an expression matrix
available in the connection object.
}

