% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrc.R
\name{check_netrc}
\alias{check_netrc}
\title{Check netrc file}
\usage{
check_netrc()
}
\value{
The name of the netrc file
}
\description{
Check that there is a netrc file with a valid entry for ImmuneSpace.
}
\details{
In order to connect to ImmuneSpace, you will need a `.netrc` file in your 
contains a `machine` name (hostname of ImmuneSpace), and `login` and 
`password`. See [here](https://www.labkey.org/wiki/home/Documentation/page.view?name=netrc) 
for more information. By default \code{RCurl} will look for the file in your
home directoty.

If no netrc is available or it is not formatted properly, \code{write_netrc}
can be used to write one. Otherwise, when specifying login and password in
\code{CreateConnection}, a temporary file will be created for that connection.
}
\examples{
try(check_netrc())

}
\seealso{
CreateConnection write_netrc
}
