% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CreateConnection}
\alias{CreateConnection}
\title{CreateConnection}
\usage{
CreateConnnection(study="SDY269")
}
\arguments{
\item{study}{\code{"character"} vector naming the study.}
}
\value{
an instance of an \code{ImmuneSpaceConnection}
}
\description{
Constructor for \code{ImmuneSpaceConnection} class
}
\details{
Instantiates and \code{ImmuneSpaceConnection} for \code{study}
}

