% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGEMatrix.R
\name{EMNames}
\alias{EMNames}
\title{Change expression-matrix sample names}
\arguments{
\item{x}{An \code{ExpressionSet}. As returned by \code{getGEMatrix}}

\item{colType}{A \code{character}. The type of column names. Valid optsions
are "expsample_accession" and "participant_id".}
}
\value{
An \code{ExpressionSet}, with updated \code{sampleNames} and
\code{phenoData}.
}
\description{
Change the sampleNames of an \code{ExpressionSet} fetched by
\code{getGEMatrix} using the information in the \code{phenodData} slot.
}

