% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConnection.R
\docType{data}
\name{ISCon}
\alias{ISCon}
\alias{ImmuneSpace}
\alias{ImmuneSpaceConnection}
\alias{ImmuneSpaceConnection-class}
\alias{getGEMatrix}
\alias{getDataset}
\alias{listDatasets}
\alias{getGEAnalysis}
\alias{listGEAnalysis}
\alias{addTrt}
\alias{EMNames}
\alias{quick_plot}
\title{The ImmuneSpaceConnection class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ISCon
}
\value{
An instance of an ImmuneSpaceConnection for a study in `labkey.url.path`
}
\description{
A connection respresents a study or a set of studies available on ImmuneSpace.
It provides function to download and display the data within these studies.
}
\details{
Uses global variables \code{labkey.url.base}, and \code{labkey.url.path}, to
access a study. \code{labkey.url.base} should be
\code{https://www.immunespace.org/}. \code{labkey.url.path} should be
\code{/Studies/studyname}, where 'studyname' is the accession number of the
study.
The ImmunespaceConnection will initialize itself, and look for a
\code{.netrc} file in \code{"~/"} the user's home directory. The
\code{.netrc} file should contain a \code{machine}, \code{login}, and
\code{password} entry to allow access to ImmuneSpace, where \code{machine} is
the host name like "www.immunespace.org".
}
\section{Fields}{

\describe{
\item{\code{study}}{A \code{character}. The study accession number. Use an empty
string ("") to create a connection at the project level.}

\item{\code{config}}{A \code{list}. Stores configuration of the connection object
such as URL, path and username.}

\item{\code{available_datasets}}{A \code{data.table}. The table of datasets available
in the connection object.}

\item{\code{data_cache}}{A \code{list}. Stores the data to avoid downloading the same
tables multiple times.}

\item{\code{constants}}{A \code{list}. Used to store information regarding
gene-expression data.}
}}

\examples{
\dontrun{
  sdy269 <- CreateConnection("SDY269")
  sdy269
}
}
\seealso{
\code{\link{CreateConnection}}
 \code{\link{ImmuneSpaceR-package}}
}
\keyword{datasets}
