% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDataset.R
\name{ImmuneSpaceConnection_getDataset}
\alias{ImmuneSpaceConnection_getDataset}
\title{Get a dataset}
\arguments{
\item{x}{A \code{character}. The name of the dataset.}

\item{original_view}{A \code{logical}. If set to TRUE, download the ImmPort
view. Else, download the default grid view. Note: Once data is cached,
changing value of this argument won't have effect on the subsequent calls
unless \code{reload} is set to 'TRUE'.}

\item{reload}{A \code{logical}. Clear the cache. If set to TRUE, download the
dataset, whether a cached version exist or not.}

\item{...}{Arguments to be passed to the underlying \code{labkey.selectRows}.}
}
\value{
a \code{data.table}
}
\description{
Downloads a dataset and cache the result in the connection object.

con$getDataset(x, original_view = FALSE, reload = FALSE, ...)
}
\details{
Returns the dataset named 'x', downloads it if it is not already cached. Note
that if additional arguments (...) are passed, the dataset will be reloaded,
even if a cached copy exist.
}
\examples{
labkey.url.base = "https://www.immunespace.org"
labkey.url.path = "/Studies/SDY269"
sdy269 <- CreateConnection("SDY269")
sdy269$getDataset("hai")
}

