% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ImmuneSpace.R
\name{ImmuneSpaceConnection_getGEAnalysis}
\alias{ImmuneSpaceConnection_getGEAnalysis}
\alias{getGEAnalysis}
\title{Get gene expression analysis}
\arguments{
\item{...}{A \code{list} of arguments to be passed to \code{labkey.selectRows}.}
}
\value{
A \code{data.table} containing the requested gene expression analysis
 results.
}
\description{
Download the result of a Gene epxression analysis experiment from the
gene_expression schema.
}

