% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ImmuneSpace.R
\name{ImmuneSpaceConnection_getGEFiles}
\alias{ImmuneSpaceConnection_getGEFiles}
\title{get Gene Expression Files}
\arguments{
\item{files}{A \code{character}. The name of the files to download}

\item{destdir}{A \code{character}. The destination directory}
}
\value{
An \code{integer} vector of the same length as the \code{files}
 parameter. See the Value section of \code{?download.file} for more
 information.
}
\description{
get Gene Expression Files
}
\details{
getGEFiles makes calls to base function \code{download.file} which
 in turn makes system calls to curl.
}
\examples{
#Downloads CEL files in the current directory
sdy269<-CreateConnection("SDY269")
sdy269$getGEFiles(c("GSM733843.CEL", "GSM733844.CEL"))
}

