% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ImmuneSpaceConnection_getGEMatrix}
\alias{ImmuneSpaceConnection_getGEMatrix}
\alias{getGEMatrix}
\title{get Gene Expression Matrix}
\arguments{
\item{x}{\code{"character"} name of the Gene Expression Matrix}
}
\value{
an \code{ExpressionSet}
}
\description{
get Gene Expression Matrix
}
\details{
Returns an `ExpressionSet` from the matrix named 'x', downloads it if it is not already cached.
}
\examples{
labkey.url.base="https://www.immunespace.org"
labkey.url.path="/Studies/SDY269"
labkey.user.email='gfinak at fhcrc.org'
sdy269<-CreateConnection("SDY269")
sdy269$getGEMatrix("TIV_2008")
}

