% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ImmuneSpace.R
\name{ImmuneSpaceConnection_listDatasets}
\alias{ImmuneSpaceConnection_listDatasets}
\alias{listDatasets}
\title{list available datasets}
\value{
Doesn't return anything, just prints to console.
}
\description{
list available datasets
}
\details{
Prints the names of the available datasets
}
\examples{
labkey.url.base="https://www.immunespace.org"
labkey.url.path="/Studies/SDY269"
labkey.user.email='gfinak at fhcrc.org'
sdy269<-CreateConnection("SDY269")
sdy269$listDatasets()
}

