% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ImmuneSpace.R
\name{ImmuneSpaceConnection_listGEAnalysis}
\alias{ImmuneSpaceConnection_listGEAnalysis}
\title{List available gene expression analysis}
\value{
A \code{data.frame}. The list of gene expression analysis.
}
\description{
List the available gene expression analysis runs.

con$listGEAnalysis()
}
\details{
Prints the table of differential expression analysis
}
\examples{
labkey.url.base="https://www.immunespace.org"
labkey.url.path="/Studies/SDY269"
sdy269<-CreateConnection("SDY269")
sdy269$listGEAnalysis()
}

