% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ImmuneSpace.R
\docType{package}
\name{ImmuneSpaceR-package}
\alias{ImmuneSpaceR}
\alias{ImmuneSpaceR-package}
\title{A Thin Wrapper Around ImmuneSpace.}
\description{
ImmuneSpaceR provides a convenient API for accessing data sets
within the ImmuneSpace database.
}
\details{
Uses the Rlabkey package to connect to ImmuneSpace. Implements
caching, and convenient methods for accessing data sets.
}
\author{
Greg Finak
}

