% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netrc.R
\name{interactive_netrc}
\alias{interactive_netrc}
\title{Interactively write a netrc file}
\usage{
interactive_netrc()
}
\value{
A netrc file that is verified to connect to ImmuneSpace
}
\description{
Write a netrc file that is valid for accessing ImmuneSpace
}
\examples{
\dontrun{
interactive_netrc()
}

}
