% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{ISpalette}
\alias{ISpalette}
\title{ImmuneSpace palette}
\usage{
ISpalette(n)
}
\arguments{
\item{n}{A \code{numeric}. The length of the desired palette.}
}
\value{
A \code{character} vector colors in hexadecimal code of length
 \code{n}.
}
\description{
Create a color gradient of the selected length that matches the ImmuneSpace
theme.
}

