% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HIPCMatrix.R
\name{isRunningOnServer}
\alias{isRunningOnServer}
\title{Detect whether the code is running on an immunespace server}
\value{
A \code{logical}. TRUE if the code is executed on one of the
immunespace servers. FALSE otherwise.
}
\description{
Detect whether the code is running on an immunespace server
}

