% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadConnection}
\alias{loadConnection}
\alias{saveConnection}
\title{Load an ImmuneSpaceConnection object from disk}
\usage{
loadConnection(file)

saveConnection(con, file)
}
\arguments{
\item{file}{the file name to be saved to or loaded from}

\item{con}{An \code{ImmuneSpaceConnection}. The connection to save to file. 
To be loaded later using \code{loadConnection}.}
}
\value{
An ImmuneSpaceConnection object
}
\description{
Load an ImmuneSpaceConnection object from disk

Save an ImmuneSpaceConnection object to disk
}

