% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPCMatrix.R
\name{makeMatrix}
\alias{makeMatrix}
\title{makeMatrix}
\arguments{
\item{gef}{A \code{data.table}. The gene_expression_files dataset, filtered
for the selected cohort and relvant files or accession numbers.}

\item{isGEO}{A \code{logical}. Set to TRUE if the Matrix should be generated
from GEO accession numbers instead of data on disk.}
}
\description{
Create a standard expression matrix from a given set of files or GEO
accession numbers.
}

