% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{quick_plot}
\alias{quick_plot}
\title{Quick plot of a data set from a study}
\usage{
quick_plot(dt, normalize_to_baseline = TRUE, type="auto", ...)
}
\arguments{
\item{dt}{\code{data.table} the data to be plotted.}

\item{normalize_to_baseline}{\code{logical} defaults to \code{TRUE}.}

\item{type}{\code{character} one of \code{c("auto","heatmap","boxplot","line")}. Heatmap chosen automatically if there are more than 5 analytes.}

\item{...}{additional arguments passed to qplot.}
}
\value{
\code{ggplot2} structure.
}
\description{
Generates a quick plot of a data set automatically. Based on ggplot2's qplot.
}
\examples{
\dontrun{
study <- CreateConnection("SDY269")
dt <- study$getDataset("elisa_mbaa")
quick_plot(dt, normalize_to_baseline = F, type="auto")
quick_plot(dt, normalize_to_baseline = F, type="boxplot")
}
}

