% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{RImmuneSpace-package}
\alias{RImmuneSpace}
\alias{RImmuneSpace-package}
\title{A Thin Wrapper Around Immunspace.}
\description{
RImmuneSpace provides a convenient API for accessing data sets within the ImmuneSpace database.
}
\details{
Uses the Rlabkey package to connect to ImmuneSpace. Implements caching, and convenient methods for accessing data sets.
}
\author{
Greg Finak
}

