% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ImmuneSpaceConnectionList.R
\name{studyNames}
\alias{studyNames}
\title{print names of the studies associated with the connection list.}
\description{
print names of the studies associated with the connection list.
}
\examples{
#\dontrun{
#cons <- CreateConnection(c("SDY269", "SDY180"))
#cons$studyNames()
#}
}

