% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{template_IS}
\alias{template_IS}
\title{template_IS}
\usage{
template_IS()
}
\arguments{
\item{None}{}
}
\value{
None
}
\description{
Template for knitted reports that matches ImmuneSpace's graphic style.
The theme among other things modifies the background and text colors, font sizes and types.
}
\details{
None.
}
\examples{
None.
}

