% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_IS}
\alias{theme_IS}
\title{theme_IS}
\usage{
theme_IS(base_size = 12)
}
\arguments{
\item{base_size}{A \code{numeric}. Base font size.}
}
\value{
A theme object
}
\description{
Theme that matches ImmuneSpace's graphic style. The theme modifies the 
background, the grid lines, the axis, and the colors used by continuous and
gradient scales.
}
\examples{
library(ggplot2)
p <- ggplot(data = mtcars) + geom_point(aes(x = mpg, y = cyl, color = hp)) + facet_grid(vs ~ am)
p + theme_IS()
}

