% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{CreateConnection}
\alias{CreateConnection}
\title{CreateConnection}
\usage{
CreateConnection(study = NULL, verbose = FALSE)
}
\arguments{
\item{study}{\code{"character"} vector naming the study.}

\item{verbose}{\code{"logical"} wehther to print the extra details for troubleshooting.}
}
\value{
an instance of an \code{ImmuneSpaceConnection} or \code{ImmuneSpaceConnectionList}
}
\description{
Constructor for \code{ImmuneSpaceConnection} class
}
\details{
Instantiates and \code{ImmuneSpaceConnection} for \code{study}
The constructor will try to take the values of the various `labkey.*` parameters from the global environment.
If they don't exist, it will use default values. These are assigned to `options`, which are then used by the \code{ImmuneSpaceConnection} class.
}

