% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConnection.R
\name{CreateConnection}
\alias{CreateConnection}
\title{CreateConnection}
\usage{
CreateConnection(study = NULL, login = NULL, password = NULL,
  use.data.frame = FALSE, verbose = FALSE)
}
\arguments{
\item{study}{A \code{"character"} vector naming the study.}

\item{login}{A \code{"character"}. Optional argument. If there is no netrc
file a temporary one can be written by passing login and password of an 
active ImmuneSpace account.}

\item{password}{A \code{"character"}. Optional. The password for the selected
login.}

\item{use.data.frame}{A \code{"logical"}. If set to TRUE, the functions will
return \code{data.frame} objects instead of \code{data.table}.}

\item{verbose}{A \code{"logical"} whether to print the extra details for 
troubleshooting.}
}
\value{
an instance of an \code{ImmuneSpaceConnection}
}
\description{
Constructor for \code{ImmuneSpaceConnection} class
}
\details{
Instantiates an \code{ImmuneSpaceConnection} for \code{study}
The constructor will try to take the values of the various `labkey.*` 
parameters from the global environment. If they don't exist, it will use 
default values. These are assigned to `options`, which are then used by the 
\code{ImmuneSpaceConnection} class.
}
\examples{
\dontrun{
  # Single study 
  con <- CreateConnection("SDY269")
  # Cross study
  con <- CreateConnection("")
}

sdy <- try(CreateConnection("SDY269"))
if(inherits(sdy, "try-error")){
  print("Read the Introduction vignette for more information on how to set up
  a .netrc file.")
}
}
\seealso{
ImmuneSpaceConnection
}
