% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ImmuneSpaceConnection}
\alias{ImmuneSpaceConnection}
\alias{ImmuneSpaceConnection-class}
\title{The ImmuneSpaceConnection class}
\value{
An instance of an ImmuneSpaceConnection for a study in `labkey.url.path`
}
\description{
Instantiate this class to access a study
}
\details{
Uses global variables \code{labkey.url.base}, and \code{labkey.url.path}, to access a study.
\code{labkey.url.base} should be \code{https://www.immunespace.org/}.
\code{labkey.url.path} should be \code{/Studies/studyname}, where 'studyname' is the name of the study.
The ImmunspaceConnection will initialize itself, and look for a \code{.netrc} file in \code{"~/"} the user's home directory.
The \code{.netrc} file should contain a \code{machine}, \code{login}, and \code{password} entry to allow access to ImmuneSpace,
where \code{machine} is the host name like "www.immunespace.org".
}
\examples{
labkey.url.base="https://www.immunespace.org"
labkey.url.path="/Studies/SDY269"
labkey.user.email='gfinak at fhcrc.org'
sdy269<-CreateConnection("SDY269")
sdy269
}
\seealso{
\code{\link{ImmuneSpaceR-package}} \code{\link{ImmuneSpaceConnection_getGEMatrix}}  \code{\link{ImmuneSpaceConnection_getDataset}}  \code{\link{ImmuneSpaceConnection_listDatasets}}
}

