% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConnection.R
\docType{class}
\name{ImmuneSpaceConnection-class}
\alias{ImmuneSpaceConnection-class}
\alias{.ISCon}
\alias{ImmuneSpace}
\alias{ImmuneSpaceConnection}
\alias{getGEMatrix}
\alias{getDataset}
\alias{listDatasets}
\alias{getGEAnalysis}
\alias{listGEAnalysis}
\alias{addTrt}
\alias{EMNames}
\alias{quick_plot}
\title{The ImmuneSpaceConnection class}
\value{
An instance of an ImmuneSpaceConnection for a study in `labkey.url.path`
}
\description{
A connection respresents a study or a set of studies available on ImmuneSpace.
It provides function to download and display the data within these studies.
}
\details{
Uses global variables \code{labkey.url.base}, and \code{labkey.url.path}, to
access a study. \code{labkey.url.base} should be
\code{https://www.immunespace.org/}. \code{labkey.url.path} should be
\code{/Studies/studyname}, where 'studyname' is the accession number of the
study.
The ImmunespaceConnection will initialize itself, and look for a
\code{.netrc} file in \code{"~/"} the user's home directory. The
\code{.netrc} file should contain a \code{machine}, \code{login}, and
\code{password} entry to allow access to ImmuneSpace, where \code{machine} is
the host name like "www.immunespace.org".
}
\section{Fields}{

\describe{
\item{\code{study}}{A \code{character}. The study accession number. Use an empty
string ("") to create a connection at the project level.}

\item{\code{config}}{A \code{list}. Stores configuration of the connection object 
such as URL, path and username.}

\item{\code{available_datasets}}{A \code{data.table}. The table of datasets available
in the connection object.}

\item{\code{data_cache}}{A \code{list}. Stores the data to avoid downloading the same
tables multiple times.}

\item{\code{constants}}{A \code{list}. Used to store information regarding 
gene-expression data.}
}}

\section{Methods}{

\describe{
\item{\code{addTrt(x = NULL)}}{Add treatment information to the phenoData of an expression matrix
     available in the connection object.

    x: A character. The name of a expression matrix that has been downloaded 
    from the connection.}

\item{\code{clear_cache()}}{Clear the data_cache. Remove downloaded datasets and expression matrices.}

\item{\code{EMNames(EM = NULL, colType = "participant_id")}}{Change the sampleNames of an ExpressionSet fetched by getGEMatrix using the
    information in the phenodData slot.

    x: An ExpressionSet, as returned by getGEMatrix.

    colType: A character. The type of column names. Valid options are 'expsample_accession'
    and 'participant_id'.}

\item{\code{getDataset(x, original_view = FALSE, reload = FALSE, colFilter = NULL,
  ...)}}{Get a dataset form the connection

    original_view: A logical. If set tot TRUE, download the ImmPort view.
    Else, download the default grid view.

    reload: A logical. Clear the cache. If set to TRUE, download the dataset,
    whether a cached version exist or not.

    colFilter: A character. A filter as returned by Rlabkey's makeFilter function.

    '...': Extra arguments to be passed to labkey.selectRows.}

\item{\code{getGEAnalysis(...)}}{Downloads data from the gene expression analysis results table.

    '...': A list of arguments to be passed to labkey.selectRows.}

\item{\code{getGEFiles(files, destdir = ".")}}{Download gene expression raw data files.

    files: A character. Filenames as shown on the gene_expression_files dataset.

    destdir: A character. The loacal path to store the downloaded files.}

\item{\code{getGEMatrix(x = NULL, cohort = NULL, summary = FALSE, reload = FALSE)}}{Downloads a normalized gene expression matrix from ImmuneSpace.

    `x': A `character'. The name of the gene expression matrix to download.

    `cohort': A `character'. The name of a cohort that has an associated gene
    expression matrix. Note that if `cohort' isn't NULL, then `x' is ignored.
    `summary': A `logical'. If set to TRUE, Downloads a matrix with expression
    averaged by gene symbol.
    `reload': A `logical'. If set to TRUE, the matrix will be downloaded again,
    even if a cached cop exist in the ImmuneSpaceConnection object.}

\item{\code{listDatasets(which = c("datasets", "expression"))}}{List the datasets available in the study or studies of the connection.}

\item{\code{listGEAnalysis()}}{List available gene expression analysis for the connection.}

\item{\code{quick_plot(...)}}{Plots a selected dataset. This is the function used by the DataExplorer 
    module on ImmuneSpace.

    dataset: A character. The name of the dataset to plot, as displayed by the
    listDataset method.

    normalize_to_baseline: A logical. If set to TRUE, the values are plotted as
    log2 fold-change from baseline.

    type: A character. The type of plot. Valid choices are 'auto', 'heatmap',
    'boxplot', 'lineplot', 'violinplot'. If set to 'auto', the function will
    select an appropriate plot type for the selected data.

    filter: A filter as created by the makeFilter function from Rlabkey.

    facet: The facetting for ggplot2 based plots. Valid choices are 'grid' and 
    'wrap'.

    text_size: The size of all text elements in the plot.

    legend: A character. Columns of the dataset or demographics to be added as
    legend on the heatmap. This argument is ignored if the plot type isn't 
    heatmap.

    show_virus_strain: A logical. Should all the virus strains be shown or should
    the values be averaged. Only used when dataset = 'hai'.

    interactive: A logical. If set to TRUE, an interactive plot will be created. 
    The default is FALSE.

    '...': Extra argument to be passed to ggplot. e.g: shape = 'Age', color =
    'Race'.}
}}

\examples{
\dontrun{
  sdy269 <- CreateConnection("SDY269")
  sdy269
}
}
\seealso{
\code{\link{CreateConnection}}
 \code{\link{ImmuneSpaceR-package}}
}
