% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConnection.R
\docType{class}
\name{ImmuneSpaceConnection}
\alias{ImmuneSpaceConnection}
\alias{ISCon}
\alias{getGEMatrix}
\alias{getDataset}
\alias{listDatasets}
\alias{getGEAnalysis}
\alias{listGEAnalysis}
\alias{addTreatmentt}
\alias{EMNames}
\alias{quick_plot}
\title{The ImmuneSpaceConnection class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ISCon
}
\value{
An instance of an ImmuneSpaceConnection for a study in
\code{labkey.url.path}.
}
\description{
A connection respresents a study or a set of studies available on ImmuneSpace.
It provides function to download and display the data within these studies.
}
\details{
Uses global variables \code{labkey.url.base}, and \code{labkey.url.path}, to
access a study. \code{labkey.url.base} should be
\code{https://www.immunespace.org/}. \code{labkey.url.path} should be
\code{/Studies/studyname}, where 'studyname' is the accession number of the
study. The ImmuneSpaceConnection will initialize itself, and look for a
\code{.netrc} file in \code{"~/"} the user's home directory. The
\code{.netrc} file should contain a \code{machine}, \code{login}, and
\code{password} entry to allow access to ImmuneSpace, where \code{machine} is
the host name like "www.immunespace.org".
}
\section{Constructor}{

\code{\link{CreateConnection}}
}

\section{Fields}{

\describe{
  \item{\code{study}}{
    A \code{character}. The study accession number. Use an empty string ("")
    to create a connection at the project level.
  }
  \item{\code{config}}{
    A \code{list}. Stores configuration of the connection object such as
    URL, path and username.
  }
  \item{\code{availableDatasets}}{
    A \code{data.table}. The table of datasets available in the connection
    object.
  }
  \item{\code{cache}}{
    A \code{list}. Stores the data to avoid downloading the same tables
    multiple times.
  }
}
}

\section{Methods}{

\describe{
  \item{\code{initialize(..., config = NULL)}}{
    Initialize \code{ImmuneSpaceConnection} class.
    See \code{\link{CreateConnection}}.
  }
  \item{\code{print()}}{
    Print \code{ImmuneSpaceConnection} class.
  }
  \item{\code{getDataset(x, original_view = FALSE, reload = FALSE,
  colFilter = NULL, ...)}}{
    Get a dataset form the connection

    \code{original_view}: A logical. If set tot TRUE, download the ImmPort
    view. Else, download the default grid view.

    \code{reload}: A logical. Clear the cache. If set to TRUE, download the
    dataset, whether a cached version exist or not.

    \code{colFilter}: A character. A filter as returned by Rlabkey's
    \code{makeFilter} function.

    \code{...}: Extra arguments to be passed to \code{labkey.selectRows}.
  }
  \item{\code{addTreatment(matrixName = NULL)}}{
    Adds treatment information to the phenoData of an expression matrix
    available in the connection object.

    \code{x}: A character. The name of a expression matrix that has been
     downloaded from the connection.
  }
  \item{\code{getGEMatrix(matrixName = NULL, cohort = NULL,
  outputType = "summary", annotation = "latest", reload = FALSE)}}{
    Downloads a normalized gene expression matrix from ImmuneSpace.

    \code{x}: A character. The name of the gene expression matrix to download.

    \code{cohort}: A character. The name of a cohort that has an associated
    gene expression matrix. Note that if `cohort` isn't NULL, then `x` is
    ignored.

    \code{outputType}: one of 'raw', 'normalized' or 'summary'. If 'raw'
    then returns an expression matrix of non-normalized values by probe.
    'normalized' returns normalized values by probe.  'summary' returns
    normalized values averaged by gene symbol.

    \code{annotation}: one of 'default', 'latest', or 'ImmSig'. Determines
    which feature annotation set is used.  'default' uses the fas from when
    the matrix was generated. latest' uses a recently updated fas based on
    the original.'ImmSig' is specific to studies involved in the
    ImmuneSignatures project and uses the annotation from when the
    meta-study's manuscript was created.

    \code{reload}: A logical. If set to TRUE, the matrix will be downloaded
    again, even if a cached cop exist in the ImmuneSpaceConnection object.
  }
  \item{\code{EMNames(EM = NULL, colType = "participant_id")}}{
    Change the sampleNames of an ExpressionSet fetched by \code{getGEMatrix}
    using the information in the phenodData slot.

    \code{x}: An ExpressionSet, as returned by \code{getGEMatrix}.

    \code{colType}: A character. The type of column names. Valid options are
    'expsample_accession' and 'participant_id'.
  }
  \item{\code{listDatasets(output = c("datasets", "expression"))}}{
    Lists the datasets available in the study or studies of the connection.
  }
  \item{\code{listGEAnalysis()}}{
    Lists available gene expression analysis for the connection.
  }
  \item{\code{listGEMatrices()}}{
    Lists available gene expression matrices for the connection.
  }
  \item{\code{getGEAnalysis(...)}}{
    Downloads data from the gene expression analysis results table.

    \code{...}: A list of arguments to be passed to \code{labkey.selectRows}.
  }
  \item{\code{clearCache()}}{
    Clears the cache. Removes downloaded datasets and expression
    matrices.
  }
  \item{\code{getGEFiles(files, destdir = ".", quiet = FALSE)}}{
    Download gene expression raw data files.

    \code{files}: A character. Filenames as shown on the
    gene_expression_files dataset.

    \code{destdir}: A character. The local path to store the downloaded
    files.
  }
  \item{\code{getGEInputs()}}{
    Downloads data from the gene expression input samples table.
  }
  \item{\code{listParticipantGroups()}}{
    Returns a dataframe with all saved Participant Groups on ImmuneSpace.
  }
  \item{\code{getParticipantData(group, dataType, original_view = FALSE,
  ...)}}{
    Returns a dataframe with ImmuneSpace data subset by groupId.

    \code{group}: Use con$listParticipantGroups() to find Participant groupId
    or groupName.

    \code{dataType}: Use \code{con$listDatasets('datasets')} to see possible
    dataType inputs.
  }
  \item{\code{quick_plot(...)}}{
    "Plots a selected dataset. This is the function used by the DataExplorer
    module on ImmuneSpace.

    \code{dataset}: A character. The name of the dataset to plot, as
    displayed by the listDataset method.

    \code{normalize_to_baseline}: A logical. If set to TRUE, the values are
    plotted as log2 fold-change from baseline.

    \code{type}: A character. The type of plot. Valid choices are 'auto',
    'heatmap', 'boxplot', 'lineplot', 'violinplot'. If set to 'auto', the
    function will select an appropriate plot type for the selected data.

    \code{filter}: A filter as created by the makeFilter function from
    Rlabkey.

    \code{facet}: The facetting for ggplot2 based plots. Valid choices are
    'grid' and 'wrap'.

    \code{text_size}: The size of all text elements in the plot.

    \code{legend}: A character. Columns of the dataset or demographics to be
    added as legend on the heatmap. This argument is ignored if the plot type
    isn't heatmap.

    \code{show_virus_strain}: A logical. Should all the virus strains be
    shown or should the values be averaged. Only used when dataset = 'hai'.

    \code{interactive}: A logical. If set to TRUE, an interactive plot will
    be created. The default is FALSE.

    \code{...}: Extra argument to be passed to ggplot. e.g: shape = 'Age',
    color = 'Race'.
  }
}
}

\examples{
\dontrun{
# Create a connection (Initiate a ImmuneSpaceConnection object)
sdy269 <- CreateConnection("SDY269")
sdy269
}

}
\seealso{
\code{\link{CreateConnection}} \code{\link{ImmuneSpaceR-package}}
}
\keyword{datasets}
