% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{ImmuneSpaceConnection_getDataset}
\alias{ImmuneSpaceConnection_getDataset}
\alias{getDataset}
\title{get a dataset}
\arguments{
\item{x}{A \code{character}. The name of the dataset}

\item{original_view}{A \code{logical}. If set to TRUE, download the ImmPort view. Else,
download the default grid view.}

\item{reload}{A \code{logical}. Clear the cache. If set to TRUE, download the
dataset, whether a cached versio exist or not.}
}
\value{
a \code{data.table}
}
\description{
get a dataset
}
\details{
Returns the dataset named 'x', downloads it if it is not already cached.
}
\examples{
labkey.url.base="https://www.immunespace.org"
labkey.url.path="/Studies/SDY269"
labkey.user.email='gfinak at fhcrc.org'
sdy269<-CreateConnection("SDY269")
sdy269$getDataset("hai")
}

