% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{ImmuneSpaceConnection_getDataset}
\alias{ImmuneSpaceConnection_getDataset}
\alias{getDataset}
\title{get a dataset}
\arguments{
\item{x}{\code{"character"} name of the dataset}
}
\value{
a \code{data.table}
}
\description{
get a dataset
}
\details{
Returns the dataset named 'x', downloads it if it is not already cached.
}
\examples{
labkey.url.base="https://www.immunespace.org"
labkey.url.path="/Studies/SDY269"
labkey.user.email='gfinak at fhcrc.org'
sdy269<-CreateConnection("SDY269")
sdy269$getDataset("hai")
}

