% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ImmuneSpace.R
\name{ImmuneSpaceConnection_getDataset}
\alias{ImmuneSpaceConnection_getDataset}
\alias{getDataset}
\title{get a dataset}
\arguments{
\item{x}{A \code{character}. The name of the dataset}

\item{original_view}{A \code{logical}. If set to TRUE, download the ImmPort
view. Else, download the default grid view. Note: Once data is cached,
changing value of this argument won't have effect on the subsequent calls
unless \code{reload} is set to 'TRUE'.}

\item{...}{Arguments to be passed to the underlying \code{labkey.selectRows}.}

\item{reload}{A \code{logical}. Clear the cache. If set to TRUE, download the
dataset, whether a cached version exist or not.}
}
\value{
a \code{data.table}
}
\description{
get a dataset
}
\details{
Returns the dataset named 'x', downloads it if it is not already cached. Note
that if additional arguments (...) are passed, the dataset will not be
reloaded.
}
\examples{
labkey.url.base="https://www.immunespace.org"
labkey.url.path="/Studies/SDY269"
labkey.user.email='gfinak at fhcrc.org'
sdy269<-CreateConnection("SDY269")
sdy269$getDataset("hai")
}

