% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ImmuneSpace.R
\name{ImmuneSpaceConnection_getGEMatrix}
\alias{ImmuneSpaceConnection_getGEMatrix}
\alias{getGEFiles}
\alias{getGEMatrix}
\title{get Gene Expression Matrix}
\arguments{
\item{x}{\code{"character"} name of the Gene Expression Matrix}

\item{files}{A \code{character}. The name of the files to download}

\item{destdir}{A \code{character}. The destination directory}
}
\value{
an \code{ExpressionSet}

An \code{integer} vector of the same length as the \code{files}
parameter. See the Value section of \code{?download.file} for more
information.
}
\description{
get Gene Expression Matrix

get Gene Expression Files
}
\details{
Returns an `ExpressionSet` from the matrix named 'x', downloads it if it is not already cached.

getGEFiles makes calls to base function \code{download.file} which
in turn makes system calls to curl.
}
\examples{
labkey.url.base="https://www.immunespace.org"
labkey.url.path="/Studies/SDY269"
labkey.user.email='gfinak at fhcrc.org'
sdy269<-CreateConnection("SDY269")
sdy269$getGEMatrix("TIV_2008")
#Downloads CEL files in the current directory
sdy269<-CreateConnection("SDY269")
sdy269$getGEFiles(c("GSM733843.CEL", "GSM733844.CEL"))
}

