% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getGEMatrix.R
\name{ImmuneSpaceConnection_getGEMatrix}
\alias{ImmuneSpaceConnection_getGEMatrix}
\alias{getGEMatrix}
\title{get Gene Expression Matrix}
\arguments{
\item{x}{A \code{character}. The name of the Gene Expression Matrix to download.}

\item{cohort}{A \code{character}. The name of a cohort that has an associated
gene expression matrix. Note that if cohort is not NULL, then x will be ignored.}

\item{summary}{A \code{logical}. If set to TRUE. Downloads a matrix with
expression averaged by gene symbol. By default the feature level matrix is
downloaded.}

\item{reload}{A \code{logical}. If set to TRUE, the matrix will be downloaded
 again, even if a cached copy exist in the \code{ImmuneSpaceConnection} object.}
}
\value{
An \code{ExpressionSet}.
}
\description{
Downloads a normalized gene expression matrix from ImmuneSpace.

con$getGEMatrix(x = NULL, cohort = NULL, summary = FALSE, reload = FALSE)
}
\examples{
labkey.url.base = "https://www.immunespace.org"
labkey.url.path = "/Studies/SDY269"
sdy269 <- CreateConnection("SDY269")
sdy269$getGEMatrix("TIV_2008")
}

