% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ImmuneSpace.R
\name{ImmuneSpaceConnection_getHAIResponse}
\alias{ImmuneSpaceConnection_getHAIResponse}
\alias{getHAIResponse}
\title{Get HAI response at peak immunogenicity}
\value{
A \code{data.table} with columns subject_accession, response and arm name
}
\description{
Peak immunogenicity is defined as the timepoint with the maximum average fold
change to baseline. It is calculated per cohort.
}

