% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ImmuneSpaceConnection_listDatasets}
\alias{ImmuneSpaceConnection_listDatasets}
\alias{listDatasets}
\title{list available datasets}
\usage{
listDatasets()
}
\value{
Doesn't return anything, just prints to console.
}
\description{
list available datasets
}
\details{
Prints the names of the available datasets
}
\examples{
labkey.url.base="https://www.immunespace.org"
labkey.url.path="/Studies/SDY269"
labkey.user.email='gfinak at fhcrc.org'
sdy269<-new("ImmuneSpaceConnection")
sdy269$listDatasets()
}

