% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ImmuneSpace.R
\name{ImmuneSpaceConnection_listDatasets}
\alias{ImmuneSpaceConnection_listDatasets}
\title{List available datasets}
\value{
Doesn't return anything, just prints to console.
}
\description{
List the datasets available in the study or studies of the connection.

con$listDatasets()
}
\details{
Prints the names of the available datasets
}
\examples{
labkey.url.base = "https://www.immunespace.org"
labkey.url.path = "/Studies/SDY269"
sdy269 <- CreateConnection("SDY269")
sdy269$listDatasets()
}

