% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ImmuneSpace.R
\name{ImmuneSpaceConnection_listGEAnalysis}
\alias{ImmuneSpaceConnection_listGEAnalysis}
\alias{listGEAnalysis}
\title{list available gene expression analysis}
\value{
A \code{data.frame}. The list of gene expression analysis.
}
\description{
list available gene expression analysis
}
\details{
Prints the table of differential expression analysis
}
\examples{
labkey.url.base="https://www.immunespace.org"
labkey.url.path="/Studies/SDY269"
labkey.user.email='gfinak at fhcrc.org'
sdy269<-CreateConnection("SDY269")
sdy269$listGEAnalysis()
}

