% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ImmuneSpaceConnectionList.R
\docType{class}
\name{ImmuneSpaceConnectionList}
\alias{.ISConList}
\alias{ImmuneSpaceConnectionList}
\alias{ImmuneSpaceConnectionList-class}
\title{The ImmuneSpaceConnectionList class}
\value{
An instance of an ImmuneSpaceConnection for a study in `labkey.url.path`
}
\description{
Instantiate this class to access a list of studies
}
\examples{
labkey.url.base <- "https://www.immunespace.org"
labkey.url.path <- "/Studies/SDY269"
labkey.user.email <- 'gfinak at fhcrc.org'
cons <- CreateConnection(c("SDY269", "SDY180"))
cons
}
\seealso{
\code{\link{ImmuneSpaceConnection-class}}
}

