% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{loadConnection}
\alias{loadConnection}
\alias{saveConnection}
\title{Load an ImmuneSpaceConnection/ImmuneSpaceConnectionList object from disk}
\usage{
loadConnection(file)

saveConnection(con, file)
}
\arguments{
\item{file}{the file name to be saved to or loaded from}
}
\value{
ImmuneSpaceConnection or ImmuneSpaceConnectionList object
}
\description{
Load an ImmuneSpaceConnection/ImmuneSpaceConnectionList object from disk

Save an ImmuneSpaceConnection/ImmuneSpaceConnectionList object to disk
}

