% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadConnection}
\alias{loadConnection}
\alias{saveConnection}
\title{Save/Load an ImmuneSpaceConnection object from disk}
\usage{
loadConnection(file)

saveConnection(con, file)
}
\arguments{
\item{file}{the file name to be saved to or loaded from}

\item{con}{An \code{ImmuneSpaceConnection}. The connection to save to file. 
To be loaded later using \code{loadConnection}.}
}
\value{
An ImmuneSpaceConnection object
}
\description{
Connection can hold a lot of data in cache. If a lot of work has 
been done (e.g: lot's of downloaded datasets and gene-expression matrices),
it can be useful to save the connection for later or offline use.
}

