% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadConnection}
\alias{loadConnection}
\alias{saveConnection}
\title{Save/Load an ImmuneSpaceConnection object from disk}
\usage{
loadConnection(file)

saveConnection(con, file)
}
\arguments{
\item{file}{The file name to be saved to or loaded from}

\item{con}{An \code{ImmuneSpaceConnection}. The connection to save to file.
To be loaded later using \code{loadConnection}.}
}
\value{
An ImmuneSpaceConnection object
}
\description{
Connection can hold a lot of data in cache. If a lot of work has
been done (e.g: lots of downloaded datasets and gene-expression matrices),
it can be useful to save the connection for later work or even offline use.
}
\examples{
#Sample saved connection with pre-downloaded expression matrices and datasets
saved <- system.file("extdata/saved_con.rds", package = "ImmuneSpaceR")
new_con <- loadConnection(saved)
new_con
names(new_con$data_cache)
\dontrun{
  saveConnection(new_con, tempfile())
}

}

