% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIPCMatrix.R
\name{makeMatrix}
\alias{makeMatrix}
\title{makeMatrix}
\arguments{
\item{gef}{A \code{data.table}. The gene_expression_files dataset, filtered
for the selected cohort and relvant files.}
}
\description{
Create a standard expression matrix from a given set of files.
}

