% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{template_IS}
\alias{template_IS}
\title{template_IS}
\usage{
template_IS(...)
}
\arguments{
\item{...}{See \code{\link[rmarkdown]{html_document}}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}
\description{
A HTML template for knitted reports that matches ImmuneSpace's graphic style.
It is based on \code{\link[rmarkdown]{html_document}} from the \pkg{rmarkdown} 
package with css, theme, and template parameters disabled.
}
\details{
See the documentation for \code{\link[rmarkdown]{html_document}} or the 
\href{http://rmarkdown.rstudio.com/html_document_format.html}{oneline documentation}
for additional details on using the html_document format.
Compared to html_document, it:
\itemize{
\item uses a custom css stylesheet
\item does not use bootstrap themes
}
}
\examples{
\dontrun{
library(ImmuneSpaceR)
rmarkdown::render("input.Rmd", template_IS())
rmarkdown::render("input.Rmd", template_IS(toc = TRUE))
}
}
