% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_IS}
\alias{theme_IS}
\title{theme_IS}
\usage{
theme_IS(base_size = 12)
}
\arguments{
\item{base_size}{A \code{numeric}. Base font size.}
}
\value{
A theme object
}
\description{
Theme that matches ImmuneSpace's graphic style. The theme modifies the
background, the grid lines, the axis, and the colors used by continuous and
gradient scales.
}
\details{
List of modified ggplot2 elements: \code{panel.background},
\code{panel.grid.major}, \code{panel.grid.minor}, \code{axis.ticks},
\code{axis.line.x}, \code{axis.line.y}, \code{plot.title}, and
\code{strip.background}.

The default \code{scale_fill_gradient}, \code{scale_fill_continuous},
\code{scale_colour_gradient} and \code{scale_colour_continous} are also
replaced by a custom scale.
}
\examples{
library(ggplot2)
p <- ggplot(data = mtcars) + geom_point(aes(x = mpg, y = cyl, color = hp)) + facet_grid(vs ~ am)
p + theme_IS()
}

